tasks/main.yml: add optional missing EPICS packages variables
This will hold missing packages required by
the epics_packages packages to be installed on
a system. This is mostrly used on Ubuntu systems
not having the same libraries required by the
Debian counterparts. This is wrong to do (installing
debian repositories in Ubuntu), but it is our only
option as of now.